backend/requirements.txt
Flask==2.1.2
requests==2.26.0
beautifulsoup4==4.11.1
lxml==4.6.3
stem==1.8.0 
geoip2==4.3.0 
pandas>=2.0.0
openpyxl>=3.1.0
flask-cors>=3.0.10
psutil>=5.9.0
pycryptodome>=3.15.0
pysocks>=1.7.1
cryptography>=37.0.0
python-dotenv>=0.20.0
waybackpy>=3.0.6
maxminddb>=2.2.0
maxminddb-geolite2>=2018.703
dnspython>=2.2.1
requests[socks]>=2.26.0
tldextract>=3.3.1
ipwhois>=1.2.0
pyOpenSSL>=22.0.0
shodan>=1.28.0
